requirements.txt
requests
requests-oauthlib
notion-client
github-heatmap
retrying
pendulum
python-dotenv
weread2notionpro
pycryptodome